+++ featuretools/tests/entityset_tests/test_es.py
@@ -924,44 +924,6 @@ def test_secondary_time_index(entityset):
             'second_ti': ['comments', 'second_ti']})
 
 
-def test_to_csv(entityset):
-    dirname = os.path.dirname(integration_data.__file__)
-    path = os.path.join(dirname, 'test')
-    p = dict(encoding='utf-8', engine='python')
-    entityset.to_csv(path, params=p, encoding='utf-8')
-    new_es = ft.EntitySet.read(path)
-    assert entityset.__eq__(new_es, deep=True)
-    shutil.rmtree(path)
-
-
-def test_to_pickle(entityset):
-    dirname = os.path.dirname(integration_data.__file__)
-    path = os.path.join(dirname, 'test')
-    entityset.to_pickle(path)
-    new_es = ft.EntitySet.read(path)
-    assert entityset.__eq__(new_es, deep=True)
-    shutil.rmtree(path)
-
-
-def test_to_parquet(entityset):
-    dirname = os.path.dirname(integration_data.__file__)
-    path = os.path.join(dirname, 'test')
-    entityset.to_parquet(path)
-    new_es = ft.EntitySet.read(path)
-    assert entityset.__eq__(new_es, deep=True)
-    shutil.rmtree(path)
-
-
-def test_to_parquet_with_lti():
-    entityset = ft.demo.load_mock_customer(return_entityset=True, random_seed=0)
-    dirname = os.path.dirname(integration_data.__file__)
-    path = os.path.join(dirname, 'test')
-    entityset.to_parquet(path)
-    new_es = ft.EntitySet.read(path)
-    assert entityset.__eq__(new_es, deep=True)
-    shutil.rmtree(path)
-
-
 def test_sizeof(entityset):
     total_size = 0
     for entity in entityset.entities:
